refactor(client): :hammer: Improve i18n scripts and naming
