+++ apps/client/package.json
@@ -88,8 +88,8 @@
     "postbuild": "mkdir -p ./.dist && cp -R .next/standalone/* ./.dist/ && cp -R .next/static ./.dist/apps/client/.next/ && cp -R ./public ./.dist/apps/client/",
     "check:types": "tsc --noEmit",
     "dev": "wait-on http://localhost:8000/health && next dev",
-    "i18n:import": "node ./scripts/i18nImport.js",
-    "i18n:export": "node ./scripts/i18nExport.js",
+    "i18n:import": "node ./scripts/import-i18n.js",
+    "i18n:export": "node ./scripts/export-i18n.js",
     "lint": "next lint",
     "lint:fix": "next lint --fix",
     "pr:prod": "gh pr create -B master-frontend -H staging-frontend -l release -t '[PROD-FRONT] Release' -b ''",
