+++ apps/client/scripts/export-i18n.js
